% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark_performance}
\alias{benchmark_performance}
\title{Benchmark Computational Performance}
\usage{
benchmark_performance(
  sample_sizes = c(100, 200, 500, 1000),
  n_replications = 100,
  n_simulations = 3
)
}
\arguments{
\item{sample_sizes}{Vector of sample sizes to test}

\item{n_replications}{Monte Carlo replications}

\item{n_simulations}{Number of simulations per size}
}
\value{
Data frame with timing results
}
\description{
Benchmark Computational Performance
}
