% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_reference.R
\name{generate_reference}
\alias{generate_reference}
\title{Generate Monte Carlo Reference Distribution}
\usage{
generate_reference(
  X_full,
  split_ratio = 0.8,
  n_replications = 500,
  alpha_mcd = 0.75,
  n_cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{X_full}{Full data matrix}

\item{split_ratio}{Training proportion}

\item{n_replications}{Number of replications}

\item{alpha_mcd}{MCD coverage parameter}

\item{n_cores}{Number of cores}

\item{verbose}{Print progress}
}
\value{
Data frame with reference values
}
\description{
Generate Monte Carlo Reference Distribution
}
